Clean it up
